// frontend/src/proto/common.ts
export interface AreaEdge {
    /**
     * @generated from protobuf field: string to_area_id = 1
     */
    toAreaId: string;
    /**
     * @generated from protobuf field: int32 travel_minutes = 2
     */
    travelMinutes: number;
}
/**
 * @generated from protobuf message lastmile.Area
 */
export interface Area {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
    /**
     * @generated from protobuf field: bool is_station = 3
     */
    isStation: boolean;
    /**
     * @generated from protobuf field: repeated lastmile.AreaEdge neighbours = 4
     */
    neighbours: AreaEdge[];
}
/**
 * @generated from protobuf message lastmile.UserId
 */
export interface UserId {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
}
// @generated message type with reflection information, may provide speed optimized methods
class AreaEdge$Type extends MessageType<AreaEdge> {
    constructor() {
        super("lastmile.AreaEdge", [
            { no: 1, name: "to_area_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "travel_minutes", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<AreaEdge>): AreaEdge {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.toAreaId = "";
        message.travelMinutes = 0;
        if (value !== undefined)
            reflectionMergePartial<AreaEdge>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: AreaEdge): AreaEdge {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string to_area_id */ 1:
                    message.toAreaId = reader.string();
                    break;
                case /* int32 travel_minutes */ 2:
                    message.travelMinutes = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: AreaEdge, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string to_area_id = 1; */
        if (message.toAreaId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.toAreaId);
        /* int32 travel_minutes = 2; */
        if (message.travelMinutes !== 0)
            writer.tag(2, WireType.Varint).int32(message.travelMinutes);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message lastmile.AreaEdge
 */
export const AreaEdge = new AreaEdge$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Area$Type extends MessageType<Area> {
    constructor() {
        super("lastmile.Area", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "is_station", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "neighbours", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => AreaEdge }
        ]);
    }
    create(value?: PartialMessage<Area>): Area {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = "";
        message.name = "";
        message.isStation = false;
        message.neighbours = [];
        if (value !== undefined)
            reflectionMergePartial<Area>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Area): Area {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string name */ 2:
                    message.name = reader.string();
                    break;
                case /* bool is_station */ 3:
                    message.isStation = reader.bool();
                    break;
                case /* repeated lastmile.AreaEdge neighbours */ 4:
                    message.neighbours.push(AreaEdge.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Area, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string name = 2; */
        if (message.name !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.name);
        /* bool is_station = 3; */
        if (message.isStation !== false)
            writer.tag(3, WireType.Varint).bool(message.isStation);
        /* repeated lastmile.AreaEdge neighbours = 4; */
        for (let i = 0; i < message.neighbours.length; i++)
            AreaEdge.internalBinaryWrite(message.neighbours[i], writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message lastmile.Area
 */
export const Area = new Area$Type();